+++ spec/models/grouping_spec.rb
@@ -45,14 +45,24 @@ describe Grouping do
     context 'hidden students' do
       let(:hidden) { create(:student, hidden: true) }
 
-      it 'cannot be invited' do
-        grouping.invite(hidden.user.user_name)
+      it 'can be invited by instructors' do
+        grouping.invite(hidden.user.user_name, StudentMembership::STATUSES[:accepted], invoked_by_instructor: true)
+        expect(grouping.memberships.count).to eq(1)
+        expect(grouping.memberships.first.role).to eq(hidden)
+      end
+
+      it 'cannot be invited by students' do
+        errors = grouping.invite(hidden.user.user_name)
+        expect(errors).not_to be_empty
+        expect(errors.first).to include('inactive')
         expect(grouping.memberships.count).to eq(0)
       end
 
-      it 'cannot be added' do
-        grouping.add_member(hidden)
-        expect(grouping.memberships.count).to eq(0)
+      it 'can be added' do
+        result = grouping.add_member(hidden)
+        expect(result).not_to be_nil
+        expect(grouping.memberships.count).to eq(1)
+        expect(grouping.memberships.first.role).to eq(hidden)
       end
     end