+++ Changelog.md
@@ -7,6 +7,7 @@
 ### ✨ New features and improvements
 - Enable test results downloads through the API (#7754)
 - Provide suggestions for partial student matching scans (#7760)
+- Allow inactive students to join groups (#7757)
 
 ### 🐛 Bug fixes